+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.3.0] - 2021-03-28
+# Added
+- Add new methods to `Bot` for listing and deleting guild slash commands
+- Exposed methods via new subcommands in main.rs
+
+# Changed
+- Removed implicit command registration from `Bot::run`
+- Changed `Bot::run` to be an async function
+- Removed `guild_id` from `Config` struct
+
 ## [0.2.2] - 2021-03-27
 # Added
 - Added example for how to run with docker
@@ -22,7 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved structopt argument parser into main.rs and export new `Config` replacement
 
 <!-- next-url -->
-[Unreleased]: https://github.com/yaahc/uwubot/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/yaahc/uwubot/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/yaahc/uwubot/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/yaahc/uwubot/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/yaahc/uwubot/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/yaahc/uwubot/releases/tag/v0.2.0